// SMSFCart/SMSCartWeb/src/app/models/Orders.ts
export class OrdersModel {
  ord_id: number;
  aid: number;
  uid: number;
  phone: string;
  itm_id: string;
  ord_date: string;
  ord_status: string;
  payment_mode: string;
  expe_dvy_date: string;
  actl_dvy: string;
  amount: number;
  shipping_charges: number;
  Total_amount: number;
  cardno: string;
  cardname: string;
  expmonth: number;
  expyear: number;
  cvv: number;
}